combined_tractor/requirements.txt
numpy
opencv-python
matplotlib
flask
pytesseract
torch
torchvision
pillow
gunicorn